requirements.txt
# Streamlit Meme Generator Requirements
# Core framework
streamlit>=1.28.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Database and authentication
supabase>=1.0.4
postgrest>=0.10.8

# AI and API integration  
groq>=0.4.1
openai>=1.3.0
requests>=2.31.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0